requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.2
Deprecated==1.2.14
dnspython==2.4.1
email-validator==2.0.0.post2
fastapi==0.100.1
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
orjson==3.9.2
pycparser==2.21
pydantic==2.1.1
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.4.0
PyGithub==1.59.0
Pygments==2.15.1
PyJWT==2.8.0
PyNaCl==1.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rich==13.4.2
shellingham==1.5.0.post1
sniffio==1.3.0
starlette==0.27.0
typer==0.9.0
typing_extensions==4.7.1
ujson==5.8.0
urllib3==2.0.4
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.15.0
